requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
PyYAML==6.0.1
APScheduler==3.10.4
python-telegram-bot==20.7
pytest==7.4.3
responses==0.24.1
